sandbox - move tests to `node` layer (#154779)
