+++ src/vs/workbench/services/keybinding/test/node/linux_de_ch.js
