+++ src/vs/workbench/services/keybinding/test/node/mac_de_ch.js
