+++ src/vs/workbench/services/keybinding/test/node/mac_en_us.js
