+++ src/vs/workbench/services/keybinding/test/node/mac_zh_hant2.txt
