+++ src/vs/workbench/services/keybinding/test/node/win_de_ch.txt
