+++ src/vs/workbench/services/keybinding/test/node/win_en_us.txt
