+++ src/vs/workbench/services/keybinding/test/node/win_por_ptb.js
