+++ src/vs/workbench/services/keybinding/test/node/win_ru.js
